pip-requirements.txt
APScheduler==3.6.3
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cryptography==2.8
future==0.18.2
idna==2.9
mysql-connector==2.2.9
pycparser==2.20
python-telegram-bot==12.3.0
pytz==2019.3
requests==2.23.0
six==1.14.0
tornado==6.0.4
tzlocal==2.0.0
urllib3==1.25.8
